+++ atcg_lib/src/Registration/Registration.cpp
@@ -5,7 +5,7 @@ namespace atcg
     Registration::Registration(const std::shared_ptr<PointCloud>& source, const std::shared_ptr<PointCloud>& target)
         :X(source->asMatrix()), Y(target->asMatrix())
     {
-        N = X.rows();
-        M = Y.rows();
+        N = static_cast<uint32_t>(X.rows());
+        M = static_cast<uint32_t>(Y.rows());
     }
 }
